requirements.txt
django==4.2.11

django-environ==0.11.2
django-cleanup==8.0.0
psycopg2-binary==2.9.9
django-tailwind==3.8.0
django-browser-reload==1.12.0

django-storages==1.14.2
django-cors-headers==4.3.0
django-ratelimit==4.1.0
dj-database-url==2.1.0
djangorestframework-simplejwt==5.3.1 # for authentication

django-payments[stripe]==2.0.0

pytz==2023.3.post1

django-phonenumber-field[phonenumberslite]==7.3.0

google-cloud-storage==2.12.0
google-api-python-client==2.126.0

whitenoise==6.6.0

pillow==10.3.0

gunicorn
